# openapi.yml
openapi: "3.0.0"
info:
  title: "IO onboarding PA API"
  x-summary: >-
    Public Administriations onboarding APIs for the IO project.
  description: |-
    The backend used by the onboarding portal for public administrations of the IO project.

    The website for the project is [io-website].
    Privacy Policy is [privacy].

    Development happens on github.


    [//]:  # (Riferimenti e url)
    [io-website]: https://io.italia.it
    [tos]: https://io.italia.it/note-legali/
    [privacy]: https://io.italia.it/privacy-policy/

  version: "0.0.1"
  x-project: io
  x-api-id: 49b844f5-2b1a-47a9-880a-ca480dac3736
  contact:
    url: https://io.italia.it
    email: teamdigitale@governo.it
  termsOfService: https://io.italia.it/note-legali/
externalDocs:
  url: https://io.italia.it
  description: IO Project website
servers:
  - url: https://api.pa-onboarding.dev.io.italia.it/
    description: Development environment
  - url: https://api.pa-onboarding.io.italia.it/
    description: Production environment
x-commons:
  # This section contains yaml anchors useful to implement common behaviors
  #  in the OAS.
  common-responses: &common-responses
    400:
      $ref: https://teamdigitale.github.io/openapi/0.0.5/definitions.yaml#/responses/400BadRequest
    429:
      $ref: https://teamdigitale.github.io/openapi/0.0.5/definitions.yaml#/responses/429TooManyRequests
    503:
      $ref: https://teamdigitale.github.io/openapi/0.0.5/definitions.yaml#/responses/503ServiceUnavailable
    default:
      $ref: https://teamdigitale.github.io/openapi/0.0.5/definitions.yaml#/responses/default
paths:
  /login:
    get:
      parameters:
        - name: entityID
          in: query
          required: true
          description: |-
            **IdP ID**.
            The id of the SPID IdP to perform the login through.
            *Note: Also a fake IdP with id `xx_tedsstenv2` is available.*
          schema:
            type: string
            enum: ["lepidaid", "infocertid", "sielteid", "namirialid", "timid", "arubaid", "posteid", "intesaid", "spiditalia", "xx_testenv2"]
          example: posteid
        - name: authLevel
          in: query
          required: true
          description: "**SPID level**. The security level of SPID authentication."
          schema:
            type: string
            enum: ["SpidL2", "SpidL3"]
          example: SpidL2
      responses:
        302:
          description: Redirection to the login page from the choosen IdP.
  /logout:
    post:
      security:
        - bearerAuth: []
      responses:
        <<: *common-responses
        200:
          description: Successful response
  /organizations:
    post:
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationRegistrationParams"
      responses:
        <<: *common-responses
        200:
          description: The registered organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /profile:
    get:
      security:
        - bearerAuth: []
      responses:
        <<: *common-responses
        200:
          description: The user profile information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    post:
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                workEmail:
                  $ref: "#/components/schemas/EmailAddress"
              required:
                - workEmail
      responses:
        <<: *common-responses
        200:
          description: The updated user profile information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /public-administrations:
    get:
      parameters:
        - name: search
          in: query
          required: true
          description: |
            **Public administration name**.
            The API responds with a list of results that match the searching words.
          schema:
            type: string
            example: comune gioiosa
      responses:
        <<: *common-responses
        200:
          description: |-
            Public Administrations matching the search criteria.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FoundAdministration'
components:
  schemas:
    EmailAddress:
      type: string
      format: email
    FiscalCode:
      type: string
      description: User's fiscal code.
      format: FiscalCode
      x-import: italia-ts-commons/lib/strings
      example: SPNDNL80R13C555X
    ArrayIndexUpTo4:
      type: integer
      minimum: -1
      maximum: 5
      exclusiveMaximum: true
    ExistingArrayIndexUpTo4:
      type: integer
      minimum: 0
      maximum: 5
      exclusiveMaximum: true
    NonEmptyString:
      type: string
      format: NonEmptyString
      x-import: italia-ts-commons/lib/strings
    OrganizationFiscalCode:
      type: string
      description: Organization's fiscal code.
      format: OrganizationFiscalCode
      x-import: italia-ts-commons/lib/strings
      example: 01234567890
    OrganizationScope:
      type: string
      enum:
        - "LOCAL"
        - "NATIONAL"
    UserRole:
      type: string
      enum:
        - "ORG_DELEGATE"
        - "ORG_MANAGER"
        - "DEVELOPER"
        - "ADMIN"
    LegalRepresentative:
      type: object
      properties:
        email:
          $ref: "#/components/schemas/EmailAddress"
        familyName:
          $ref: "#/components/schemas/NonEmptyString"
        firstName:
          $ref: "#/components/schemas/NonEmptyString"
        fiscalCode:
          $ref: "#/components/schemas/FiscalCode"
        phoneNumber:
          $ref: "#/components/schemas/NonEmptyString"
        role:
          $ref: "#/components/schemas/UserRole"
      required:
        - email
        - familyName
        - firstName
        - fiscalCode
        - phoneNumber
        - role
      example:
        {
          "familyName": "Spano'",
          "firstName": "Ignazio Alfonso",
          "fiscalCode": "BCDFGH12A21Z123D",
          "phoneNumber": "5550000000",
          "email": "indirizzo00@email.pec.it",
          "role": "ORG_MANAGER"
        }
    UserProfile:
      type: object
      properties:
        email:
          $ref: "#/components/schemas/EmailAddress"
        familyName:
          type: string
        firstName:
          type: string
        fiscalCode:
          $ref: "#/components/schemas/FiscalCode"
        role:
          $ref: "#/components/schemas/UserRole"
        workEmail:
          $ref: "#/components/schemas/EmailAddress"
      required:
        - email
        - familyName
        - firstName
        - fiscalCode
        - role
      example:
        {
          email: "example@email.com",
          familyName: "Rossi",
          firstName: "Mario",
          fiscalCode: "RSSMRA80A01H501U",
          role: "ORG_DELEGATE",
          workEmail: "work-email@email.com"
        }
    Link:
      externalDocs:
        url: https://tools.ietf.org/id/draft-handrews-json-schema-hyperschema-02.html
        description: HyperSchema internet draft.
      description: |-
        As this API wants to drive the consumer in the workflow, we decided to
        use Link Descriptor Objects defined in the current [json-schema draft](https://json-schema.org/draft/2019-09/links).

        [IANA link relations](https://www.iana.org/assignments/link-relations/link-relations.xhtml)
        should be used in `rel`.
      type: object
      properties:
        rel:
          type: string
        href:
          type: string
          format: uri-template
        title:
          type: string
        description:
          type: string
        $comment:
          type: string
      required:
        - rel
        - href
    Links:
      type: array
      items:
        $ref: '#/components/schemas/Link'
      example:
        - rel: self
          href: https://api.pa-onboarding.dev.io.italia.it/public-administrations/c_e956
          description: Agency URL

    FoundAdministration:
      oneOf:
        - $ref: '#/components/schemas/FoundNotRegisteredAdministration'
        - $ref: '#/components/schemas/FoundRegisteredAdministration'
    FoundNotRegisteredAdministration:
      type: object
      properties:
        fiscalCode:
          $ref: "#/components/schemas/OrganizationFiscalCode"
        ipaCode:
          type: string
        name:
          type: string
        legalRepresentative:
          type: object
          properties:
            familyName:
              type: string
            firstName:
              type: string
          required:
            - familyName
            - firstName
        links:
          $ref: '#/components/schemas/Links'
        pecs:
          type: array
          items:
            $ref: '#/components/schemas/EmailAddress'
          minLength: 1
          maxLength: 5
      required:
        - fiscalCode
        - ipaCode
        - name
        - legalRepresentative
        - link
        - pecs
      example:
        [
        {
          "fiscalCode": "86000470830",
          "ipaCode": "c_e043",
          "legalRepresentative": {
            "familyName": "Spano'",
            "firstName": "Ignazio Alfonso"
          },
          "name": "Comune di Gioiosa Marea",
          "pecs": [
            "indirizzo00@email.pec.it",
            "indirizzo01@email.pec.it"
          ],
          "links": [{
            "rel": "self",
            "href": "https://api.pa-onboarding.dev.io.italia.it/public-administrations/c_e043"
          }]
        },
        {
          "fiscalCode": "81000930800",
          "ipaCode": "c_e044",
          "legalRepresentative": {
            "familyName": "ALI'",
            "firstName": "Gianfranco"
          },
          "name": "Comune di Gioiosa Jonica",
          "pecs": [
            "indirizzo10@email.pec.it",
            "indirizzo11@email.pec.it",
            "indirizzo12@email.pec.it",
            "indirizzo13@email.pec.it",
          ],
          "links": [{
            "rel": "self",
            "href": "https://api.pa-onboarding.dev.io.italia.it/public-administrations/c_e044"
          }]
        },
        {
          "fiscalCode": "00282520808",
          "ipaCode": "c_e956",
          "legalRepresentative": {
            "familyName": "Sergio",
            "firstName": "Mazzia"
          },
          "name": "Comune di Marina di Gioiosa Ionica",
          "pecs": [
            "indirizzo20@email.pec.it",
            "indirizzo21@email.pec.it",
          ],
          "links": [{
            "rel": "self",
            "href": "https://api.pa-onboarding.dev.io.italia.it/public-administrations/c_e0956"
          }]
        }
        ]
    FoundRegisteredAdministration:
      type: object
      properties:
        fiscalCode:
          $ref: "#/components/schemas/OrganizationFiscalCode"
        ipaCode:
          type: string
        name:
          type: string
        legalRepresentative:
          type: object
          properties:
            familyName:
              type: string
            firstName:
              type: string
            fiscalCode:
              $ref: "#/components/schemas/FiscalCode"
            phoneNumber:
              type: string
          required:
            - familyName
            - firstName
            - fiscalCode
            - phoneNumber
        links:
          $ref: '#/components/schemas/Links'
        pecs:
          type: array
          items:
            $ref: '#/components/schemas/EmailAddress'
          minLength: 1
          maxLength: 5
        scope:
          $ref: '#/components/schemas/OrganizationScope'
        selectedPecIndex:
          $ref: '#/components/schemas/ArrayIndexUpTo4'
      required:
        - fiscalCode
        - ipaCode
        - name
        - legalRepresentative
        - link
        - pecs
        - scope
        - selectedPecIndex
      example:
        [
        {
          "fiscalCode": "86000470830",
          "ipaCode": "c_e043",
          "legalRepresentative": {
            "familyName": "Spano'",
            "firstName": "Ignazio Alfonso",
            "fiscalCode": "BCDFGH12A21Z123D",
            "phoneNumber": "3331111111"
          },
          "name": "Comune di Gioiosa Marea",
          "pecs": [
            "indirizzo00@email.pec.it",
            "indirizzo01@email.pec.it"
          ],
          "scope": "NATIONAL",
          "selectedPecIndex": 0,
          "links": [{
            "rel": "self",
            "href": "https://api.pa-onboarding.dev.io.italia.it/public-administrations/c_e043"
          }]
        },
        {
          "fiscalCode": "81000930800",
          "ipaCode": "c_e044",
          "legalRepresentative": {
            "familyName": "ALI'",
            "firstName": "Gianfranco",
            "fiscalCode": "BCDFGH12A21Z123F",
            "phoneNumber": "3332222222"
          },
          "name": "Comune di Gioiosa Jonica",
          "pecs": [
            "indirizzo10@email.pec.it",
            "indirizzo11@email.pec.it",
            "indirizzo12@email.pec.it",
            "indirizzo13@email.pec.it",
          ],
          "scope": "LOCAL",
          "selectedPecIndex": 2,
          "links": [{
            "rel": "self",
            "href": "https://api.pa-onboarding.dev.io.italia.it/public-administrations/c_e044"
          }]
        },
        {
          "fiscalCode": "00282520808",
          "ipaCode": "c_e956",
          "legalRepresentative": {
            "familyName": "Sergio",
            "firstName": "Mazzia",
            "fiscalCode": "BCDFGH12A21Z123E",
            "phoneNumber": "3333333333"
          },
          "name": "Comune di Marina di Gioiosa Ionica",
          "pecs": [
            "indirizzo20@email.pec.it",
            "indirizzo21@email.pec.it",
          ],
          "scope": "LOCAL",
          "selectedPecIndex": 0,
          "links": [{
            "rel": "self",
            "href": "https://api.pa-onboarding.dev.io.italia.it/public-administrations/c_e0956"
          }]
        }
        ]
    Organization:
      type: object
      properties:
        fiscalCode:
          $ref: "#/components/schemas/OrganizationFiscalCode"
        ipaCode:
          $ref: "#/components/schemas/NonEmptyString"
        legalRepresentative:
          $ref: "#/components/schemas/LegalRepresentative"
        links:
          $ref: "#/components/schemas/Links"
        name:
          $ref: "#/components/schemas/NonEmptyString"
        pec:
          $ref: "#/components/schemas/EmailAddress"
        scope:
          $ref: '#/components/schemas/OrganizationScope'
      required:
        - fiscalCode
        - ipaCode
        - name
        - pec
        - scope
      example:
        {
          "fiscalCode": "86000470830",
          "ipaCode": "c_e043",
          "legalRepresentative": {
            "familyName": "Spano'",
            "firstName": "Ignazio Alfonso",
            "fiscalCode": "BCDFGH12A21Z123D",
            "phoneNumber": "5550000000",
            "email": "indirizzo00@email.pec.it",
            "role": "ORG_MANAGER"
          },
          "name": "Comune di Gioiosa Marea",
          "pec": "indirizzo00@email.pec.it",
          "scope": "NATIONAL"
        }
    OrganizationRegistrationParams:
      type: object
      properties:
        ipaCode:
          $ref: "#/components/schemas/NonEmptyString"
        legalRepresentative:
          type: object
          properties:
            familyName:
              $ref: "#/components/schemas/NonEmptyString"
            firstName:
              $ref: "#/components/schemas/NonEmptyString"
            fiscalCode:
              $ref: "#/components/schemas/FiscalCode"
            phoneNumber:
              $ref: "#/components/schemas/NonEmptyString"
          required:
            - familyName
            - firstName
            - fiscalCode
            - phoneNumber
        scope:
          $ref: '#/components/schemas/OrganizationScope'
        selectedPecIndex:
          $ref: '#/components/schemas/ExistingArrayIndexUpTo4'
      required:
        - ipaCode
        - legalRepresentative
        - scope
        - selectedPecIndex
      example:
        {
          "ipaCode": "c_e043",
          "legalRepresentative": {
            "familyName": "Spano'",
            "firstName": "Ignazio Alfonso",
            "fiscalCode": "BCDFGH12A21Z123D",
            "phoneNumber": "3331111111"
          },
          "scope": "NATIONAL",
          "selectedPecIndex": 0
        }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer